Move test utilities to componentdocs
