+++ component/componentdocs/component_docs.go
@@ -14,7 +14,7 @@
 
 // Package componenttest define types and functions used to help test packages
 // implementing the component package interfaces.
-package componenttest
+package componentdocs
 
 import (
 	"fmt"
